hot-fix: dday 구하는 로직 변경

+++ src/main/kotlin/com/yourssu/balanssu/domain/model/dto/ViewCategoriesDto.kt
@@ -6,6 +6,6 @@ class ViewCategoriesDto(
     val categoryId: String,
     val title: String,
     val type: CategoryType,
-    val dDay: Int,
+    val dDay: Long,
     val participantCount: Int
 )
